backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
pydantic-settings==2.9.1
psutil==7.1.0
Pillow==11.2.1
safetensors==0.5.3
PyYAML==6.0.2
torch==2.4.1
diffusers==0.33.1
transformers==4.46.3
accelerate==1.5.2
peft==0.14.0
controlnet-aux==0.0.10
opencv-python==4.10.0.84
ip-adapter==0.1.0
